+++ migrations/app/20170508160706_app_custody_tiers.js
@@ -2,7 +2,7 @@
 exports.up = function (knex, Promise) {
   return knex.schema.createTable('custody_tiers', function (table) {
     table.increments('id')
-    table.bigInteger('workload_id').unsigned().notNullable().references('workload.id')
+    table.integer('workload_id').unsigned().notNullable().references('workload.id')
     table.integer('tier_type').unsigned()
     table.string('case_type', 5)
     table.integer('points').unsigned()
